+++ projects/js-packages/components/components/text/index.tsx
@@ -38,6 +38,7 @@ const Text = forwardRef< HTMLElement, TextProps >(
 		);
 	}
 );
+Text.displayName = 'Text';
 
 export default Text;