+++ projects/packages/import/src/endpoints/class-custom-css.php
@@ -7,15 +7,22 @@
 
 namespace Automattic\Jetpack\Import\Endpoints;
 
+require_once ABSPATH . 'wp-includes/theme.php';
+
 /**
  * Class Custom_CSS
  */
 class Custom_CSS extends \WP_REST_Posts_Controller {
 
 	/**
-	 * The Import ID add a new item to the schema.
+	 * Base class
 	 */
 	use Import;
+
+	/**
+	 * The Import ID add a new item to the schema.
+	 */
+	use Import_ID;
 
 	/**
 	 * Whether the controller supports batching.
@@ -34,29 +41,12 @@ class Custom_CSS extends \WP_REST_Posts_Controller {
 	}
 
 	/**
-	 * Registers the routes for the objects of the controller.
-	 *
-	 * @see WP_REST_Posts_Controller::register_rest_route()
-	 */
-	public function register_routes() {
-		register_rest_route(
-			self::$rest_namespace,
-			'/' . $this->rest_base,
-			$this->get_route_options()
-		);
-	}
-
-	/**
 	 * Update the custom CSS post.
 	 *
 	 * @param WP_REST_Request $request Full details about the request.
 	 * @return WP_REST_Response|WP_Error Response object on success, or WP_Error object on failure.
 	 */
 	public function create_item( $request ) {
-		if ( ! function_exists( 'wp_update_custom_css_post' ) ) {
-			require_once ABSPATH . 'wp-includes/theme.php';
-		}
-
 		$args = array(
 			'stylesheet' => $request['title'],
 		);