+++ projects/packages/videopress/src/client/components/timestamp-control/types.ts
@@ -1,8 +1,14 @@
+/*
+ * Types
+ */
+import type { ReactNode } from 'react';
+
 export type DecimalPlacesProp = 1 | 2 | 3;
 
 export type TimestampInputProps = {
 	disabled?: boolean;
 	value: number;
+	min?: number;
 	max?: number;
 	fineAdjustment?: number;
 	onChange?: ( ms: number ) => void;
@@ -11,6 +17,10 @@ export type TimestampInputProps = {
 };
 
 export type TimestampControlProps = TimestampInputProps & {
+	label?: ReactNode;
+	help?: ReactNode;
 	wait?: number;
+	marksEvery?: number;
+	renderTooltip?: boolean;
 	onDebounceChange?: ( ms: number ) => void;
 };