+++ projects/plugins/jetpack/extensions/store/media-source/test/index.js
@@ -5,19 +5,33 @@ import storeDefinition from '../store-definition';
 const { actions } = storeDefinition;
 const STORE_ID = 'jetpack/media-source';
 
-const setup = () => registerStore( STORE_ID, storeDefinition );
+const store = registerStore( STORE_ID, storeDefinition );
+
+// Infrastructure to clean up the media sources after each test.
+//  - Use a special function to register sources so they can be removed.
+//  - Reset the default to null after each test too.
+let mediaSourcesToCleanUp = [];
+const registerMediaSourceForTest = ( id, mediaSourceState ) => {
+	store.dispatch( actions.registerMediaSource( id, mediaSourceState ) );
+	mediaSourcesToCleanUp.push( id );
+};
+afterEach( () => {
+	for ( const id of mediaSourcesToCleanUp ) {
+		store.dispatch( actions.unregisterMediaSource( id ) );
+	}
+	mediaSourcesToCleanUp = [];
+	store.dispatch( actions.setDefaultMediaSource( null ) );
+} );
 
 describe( 'save', () => {
 	test( 'Initial State', () => {
-		const store = setup();
 		const got = store.getState();
 		const want = { sources: {}, default: null };
 		expect( got ).toEqual( want );
 	} );
 
 	test( 'Add a player with default properties', () => {
-		const store = setup();
-		store.dispatch( actions.registerMediaSource( 100, {} ) );
+		registerMediaSourceForTest( 100, {} );
 		const got = store.getState();
 		const want = {
 			default: null,
@@ -29,8 +43,7 @@ describe( 'save', () => {
 	} );
 
 	test( 'Add a player with one overriden property', () => {
-		const store = setup();
-		store.dispatch( actions.registerMediaSource( 100, { status: 'playing' } ) );
+		registerMediaSourceForTest( 100, { status: 'playing' } );
 		const got = store.getState();
 		const want = {
 			default: null,
@@ -42,13 +55,12 @@ describe( 'save', () => {
 	} );
 
 	test( 'Add two sources, then update the first one', () => {
-		const store = setup();
-		store.dispatch( actions.registerMediaSource( 100, { status: 'playing' } ) );
+		registerMediaSourceForTest( 100, { status: 'playing' } );
 		const stateAfterOneAction = store.getState();
 		const frozenStateAfterOneAction = JSON.parse( JSON.stringify( stateAfterOneAction ) );
 
-		store.dispatch( actions.registerMediaSource( 200, { status: 'stopped', position: 2 } ) );
-		store.dispatch( actions.registerMediaSource( 100, { status: 'stopped', position: 1 } ) );
+		registerMediaSourceForTest( 200, { status: 'stopped', position: 2 } );
+		registerMediaSourceForTest( 100, { status: 'stopped', position: 1 } );
 		const got = store.getState();
 		const want = {
 			default: null,
@@ -62,12 +74,11 @@ describe( 'save', () => {
 	} );
 
 	test( 'Add two sources, then delete one', () => {
-		const store = setup();
-		store.dispatch( actions.registerMediaSource( 100, { status: 'playing' } ) );
+		registerMediaSourceForTest( 100, { status: 'playing' } );
 		const stateAfterOneAction = store.getState();
 		const frozenStateAfterOneAction = JSON.parse( JSON.stringify( stateAfterOneAction ) );
 
-		store.dispatch( actions.registerMediaSource( 200, { status: 'stopped' } ) );
+		registerMediaSourceForTest( 200, { status: 'stopped' } );
 		store.dispatch( actions.unregisterMediaSource( 100 ) );
 		const got = store.getState();
 		const want = {
@@ -81,7 +92,6 @@ describe( 'save', () => {
 	} );
 
 	test( 'Set Default', () => {
-		const store = setup();
 		let got = store.getState();
 		let want = { default: null, sources: {} };
 		expect( got ).toEqual( want );
@@ -103,10 +113,8 @@ describe( 'save', () => {
 	} );
 
 	test( 'Error', () => {
-		const store = setup();
-
 		// Create sources
-		store.dispatch( actions.registerMediaSource( 100, {} ) );
+		registerMediaSourceForTest( 100, {} );
 
 		let got = store.getState();
 		let want = {
@@ -130,10 +138,8 @@ describe( 'save', () => {
 	} );
 
 	test( 'Set Time', () => {
-		const store = setup();
-
 		// Create sources
-		store.dispatch( actions.registerMediaSource( 100, {} ) );
+		registerMediaSourceForTest( 100, {} );
 		const stateAfterOneAction = store.getState();
 		const frozenStateAfterOneAction = JSON.parse( JSON.stringify( stateAfterOneAction ) );
 
@@ -169,10 +175,8 @@ describe( 'save', () => {
 	} );
 
 	test( 'Play, Pause, Toggle', () => {
-		const store = setup();
-
 		// Create sources
-		store.dispatch( actions.registerMediaSource( 100, {} ) );
+		registerMediaSourceForTest( 100, {} );
 		const stateAfterOneAction = store.getState();
 		const frozenStateAfterOneAction = JSON.parse( JSON.stringify( stateAfterOneAction ) );
 
@@ -232,10 +236,8 @@ describe( 'save', () => {
 	} );
 
 	test( 'Toggle a player with no state', () => {
-		const store = setup();
-
 		// Create sources
-		store.dispatch( actions.registerMediaSource( 100, {} ) );
+		registerMediaSourceForTest( 100, {} );
 		const stateAfterOneAction = store.getState();
 		const frozenStateAfterOneAction = JSON.parse( JSON.stringify( stateAfterOneAction ) );