+++ projects/plugins/mu-wpcom-plugin/CHANGELOG.md
@@ -5,10 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.0.11 - 2023-03-20
+## 1.1.2 - 2023-04-17
 ### Changed
 - Internal updates.
 
+## 1.1.1 - 2023-04-10
+### Changed
+- Internal updates.
+
+## 1.1.0 - 2023-04-05
+### Added
+- Launchpad: Update composer.lock [#29841]
+
+### Changed
+- Updated package dependencies. [#29565]
+
+## 1.0.12 - 2023-03-28
+
+## 1.0.11 - 2023-03-20
+### Changed
+- Internal updates.
 
 ## 1.0.10 - 2023-03-13
 ### Added